indexer/requirements.txt
# Azure SDK dependencies
azure-identity==1.17.1
azure-cosmos==4.7.0
azure-storage-blob==12.22.0
azure-ai-documentintelligence==1.0.0
azure-search-documents==11.6.0
azure-core==1.30.2

# SharePoint and Office 365
Office365-REST-Python-Client==2.5.12
msal==1.30.0

# Gremlin for graph database
gremlinpython==3.7.2

# Machine Learning and NLP
numpy==1.26.4
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
sentence-transformers==2.7.0

# Data processing and manipulation
pandas==2.2.2
requests==2.32.3
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5

# Web framework and async
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiofiles==23.2.1
asyncio-throttle==1.0.2

# Configuration and environment
pydantic==2.7.4
pydantic-settings==2.4.0
python-dotenv==1.0.1

# Logging and monitoring
structlog==24.4.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-azure-monitor==1.0.0b28

# Utilities
click==8.1.7
tenacity==8.5.0
tqdm==4.66.4
python-dateutil==2.9.0
pytz==2024.1

# Testing (optional for development)
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
httpx==0.27.0

# Development tools
black==24.8.0
isort==5.13.2
mypy==1.11.1
